+++ src/grepros/plugins/auto/postgres.py
@@ -8,7 +8,7 @@ Released under the BSD License.
 
 @author      Erki Suurjaak
 @created     02.12.2021
-@modified    04.02.2022
+@modified    27.03.2023
 ------------------------------------------------------------------------------
 """
 ## @namespace grepros.plugins.auto.postgres
@@ -161,7 +161,7 @@ class PostgresSink(DataSinkBase):
     def _make_column_value(self, value, typename=None):
         """Returns column value suitable for inserting to database."""
         TYPES = self._get_dialect_option("types")
-        plaintype = typename and rosapi.scalar(typename)  # "string<=10" -> "string"
+        plaintype = rosapi.scalar(typename)  # "string<=10" -> "string"
         v = value
         # Common in JSON but disallowed in Postgres
         replace = {float("inf"): None, float("-inf"): None, float("nan"): None}
@@ -175,7 +175,8 @@ class PostgresSink(DataSinkBase):
                 if self._nesting: v = None
                 else: v = psycopg2.extras.Json([rosapi.message_to_dict(m, replace)
                                                 for m in v], json.dumps)
-            elif "BYTEA" == TYPES.get(typename):
+            elif "BYTEA" in (TYPES.get(typename),
+                             TYPES.get(rosapi.canonical(typename, unbounded=True))):
                 v = psycopg2.Binary(bytes(bytearray(v)))  # Py2/Py3 compatible
             else:
                 v = list(self._convert_column_value(v, typename))  # Ensure not-tuple for psycopg2