+++ CHANGELOG.rst
@@ -7,6 +7,8 @@ Changelog for package grepros
 * add nesting=array|all to --write Parquet options
 * add idgenerator=callable to --write Parquet options
 * add rosapi.canonical()
+* match bounded array fields to configured output types properly 
+  in Parquet/Postgres/SQL/SQLite output, like "uint8[10]" for "BYTEA" in Postgres
 * workaround for ROS1 time/duration fields defined as int32 while actually being uint32
 * fix date formatting in HTML output